+++ electron/preload.js
@@ -5,15 +5,20 @@ contextBridge.exposeInMainWorld('electronAPI', {
   openDeckWindow: () => ipcRenderer.invoke('open-deck-window'),
   closeDeckWindow: () => ipcRenderer.invoke('close-deck-window'),
   
-  // Auto-updater events
+  // Enhanced Auto-updater API
+  onUpdateChecking: (callback) => ipcRenderer.on('update-checking', callback),
   onUpdateAvailable: (callback) => ipcRenderer.on('update-available', callback),
-  onDownloadProgress: (callback) => ipcRenderer.on('download-progress', callback),
-  onUpdateDownloaded: (callback) => ipcRenderer.on('update-downloaded', callback),
-  onUpdateError: (callback) => ipcRenderer.on('update-error', callback),
   onUpdateNotAvailable: (callback) => ipcRenderer.on('update-not-available', callback),
+  onUpdateError: (callback) => ipcRenderer.on('update-error', callback),
+  onDownloadProgress: (callback) => ipcRenderer.on('update-download-progress', callback),
+  onUpdateDownloaded: (callback) => ipcRenderer.on('update-downloaded', callback),
   
-  // Manual update check
+  // Manual update controls
   checkForUpdates: () => ipcRenderer.invoke('check-for-updates'),
+  downloadUpdate: () => ipcRenderer.invoke('download-update'),
+  installUpdate: () => ipcRenderer.invoke('install-update'),
+  getAppVersion: () => ipcRenderer.invoke('get-app-version'),
+  openReleaseNotes: (version) => ipcRenderer.invoke('open-release-notes', version),
 
   // Remove listeners
   removeAllListeners: (channel) => ipcRenderer.removeAllListeners(channel)